+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './Cipher';
 export * from './Enc';
+export * from './encryptionData';
 export * from './cbc';
 export * from './gcm';